[RFC] convert: fix --datesort
The problem is that previously commit.date was used for sorting, but it's a
string like "1 Jan xxx 2007", so it it wrong to use it for sorting.

Another problem is that why we are using depth for sorting -- I have no clear
answer -- it seems to be plain wrong.

This patch is just an RFC.

--HG--
extra : transplant_source : %B9%40%26%0CB%91%F5%B8Ed%8DDJ%99%0B%E7%A6%EB%1C%1F