add a fix for issue 1175
If we copy a file followed by an update, it's possible for the parent
manifest to no longer contain the source file of the copy, which could cause
commit to fail. If this happens, we search backwares from the first
parent to find the most likely original revision.

--HG--
extra : rebase_source : 4ea097bb15ae8b70c8d151619e8041ba04aa5dc9