py3kcompat: added fake ord implementation for py3k
In py3k, a bytes object __getitem__ will return an int instead of a
one-character bytes object. This has negative consequences when we want to
ord(), like in the following example:

>>> b'foo'[0]
102
>>> ord(b'foo'[0])
Traceback (most recent call last):
  File "<stdin>", line 1, in <module>
TypeError: ord() expected string of length 1, but int found

This patch overrides the default ord() implementation to just return an int
that's what is passed as an argument for ord(). Making the above call succeed:

>>> ord(b'foo'[0])
102