hgweb: be consistent with the background highlighting on closed branches (Issue2386)
The coal and paper web themes will highlight closed branches as a visual cue to
the user. This patch makes the other two themes constistent.

For users with difficulty differentiating colors, it would be better to split
out closed branches to a new table or optionally hide them.

--HG--
branch : stable