building: build inotify for sys.platform='linux*'
If Python interpreter was built under Linux 3.x kernel, it reports
sys.platform to be 'linux3' (it is fixed for Python 3, but not for 2.x).
This cancels building inotify extension, which was built only for 'linux2'
platform. Improved test checks if sys.platform begins with 'linux', and together
with test for kernel version to be greater than 2.6 it seems to cover all known
cases.

--HG--
branch : stable
extra : transplant_source : %0DOn%84%3B%05%A7%D8%AE%E0%08hcF%9D%C5%9B%8Bq%97