docs: don't use :hg: at the beginning of lines in notes (issue3397)
It seem like docutils 0.8 interpret ':hg:`command`' roles at the beginning of
indented lines in '.. note::' directives as a field that is an invalid argument
to the directive. It fails with 'Error in "note" directive: invalid option
block.' Docutils 0.7 accepted this arguably incorrect markup.

Reflowing the text makes the problem go away. A leading '\ ' could perhaps also
be used to mask the problem.

--HG--
branch : stable