tests: don't use dates before epoch in test-keyword.t
Timezone offsets of less than a minute is not shown but can cause displayed
dates to be before epoch start - and dates before epoch start is not shown
correctly on Windows (see also 4d5b12a5517b).

These 'negative' dates could be considered undefined behaviour so we don't care
and swap the tests values for timestamp and timezone.