hidden: drop cache on hiddenrevs property
The `filteredrevs` function already have a cache mechanism. And this cache in
invalidated at the same time than the current property cache. So we drop the
cache on the property.

The property itself is going to be dropped soon.

--HG--
extra : rebase_source : 585d2fbfbc2c5eb39b1c31193763fd581c41fef0