unbundle: extract checkheads in its own function
We are going to refactor the unbundle function to have it working on
a local repository too. Having this function extracted will ease the process.

In the case of non-matching heads, the function now directly raises an
exception. The top level of the function is catching it.

--HG--
extra : amend_source : e8b1ba3891b29491103174668680aea9d2b0e3b3