bundle2: introduce `replycaps` part for on-demand reply
The bundle2 processing does not create a bundle2 reply by default anymore.  It
is only done if the client requests it with a `replycaps` part. This part is
called `replycaps` as it will eventually contain data about which bundle2
capabilities are supported by the client.

We have to add a flag to the test command to control whether a reply is
generated or not.

--HG--
extra : rebase_source : 3cabdd0806ccbe4e89323c8f40ed145528a29778
extra : amend_source : 3399b604760e7afca4944f0c9d81ff6ee2c28a2c