subrepo: add shortid() method to subrepo classes
This method takes an "id" (e.g. a revision id) and returns a "short" version
(e.g. a short revision id).

This will be used on the next revision to fix a small bug in the way that the
text on the promptchoice shown when a subrepo diverges is generated.

--HG--
extra : rebase_source : b6d20e942c07e0bcf47a8ae502628004e7038f69