revert: call status against revert target too
We now call status against the target (and possibly against the working
directory parent is different). We do not use the information from the two
sources yet, but this is coming soon.

We need the status information aganst the dirstate in all case because we need
to be able to backup local modification.

--HG--
extra : rebase_source : 4be3caa25ac2eedda7f703ee1900b83708244920