repoview: add caching bits
Add a caching infrastructure to cache hidden changesets. The cache tries to read
the cache lazily and falls back to recomputing if no wlock can be obtain.
To validate the cache we store a sha of the obstore content and repo heads in
the beginning of the cache which we check every request.

--HG--
extra : rebase_source : 265e10e196c4d11ce72eaef17d93a1a9f0ab5df3