config: propose some sample global config file
An example of what could be suggested to the user as a global config
file. Trying to be conservative here, and only suggesting the safest
possible extensions. In addition to the user-level extensions, the
blackbox extension is something a sysadmin might reasonable want to
enable for every repo on the system.

--HG--
extra : rebase_source : d3f4f3cec62ac241ad3f6ee0edb6440785087e02