revert: have an explicit action for "forget"
The distinction between "remove" and "forget" used to be in special logic
checking for the state of the file in the dirstate. Now that we have dedicated
filtering, we can stop relying on this logic and have two distinct actions.

--HG--
extra : rebase_source : b4c7e3918bf2c58d548ab0ce509d247ff3cb166c