transaction: work around and document issue with file backup
The backup restoration is actually hard-coded for the main file. And this
hard-coded list is the only one used when repairing an interrupted transaction
from another process.

Solving this problem is out of the scope of this series so we document it and
work around its implications.

--HG--
extra : rebase_source : 1e373b2e67f61e336e45ad2383a4d912ed57a907