localrepo: add ignoremissing parameter to branchtip
Previously, in the namespaces api, the only caller of branchtip was singlenode
which happened to raise the same exception that branchtip raised: KeyError.

This is a minor change but will allow upcoming patches to use repo.branchtip to
not raise an exception if a branch doesn't exist. After that, it will be
possible for extensions to use the namespace api in a stable way.

--HG--
extra : rebase_source : d2a8e55d4b08759ec5cee026525efc56eff8afcd
extra : histedit_source : b4bf29540c5e5ca908bda9ca5b7863344534c29e