test-status-rev: remove duplicate tests
The second group of tests in test-status-rev compare to an empty
revision. The first group of tests that compare to the first commit
should be testing all the same states with the missing_* files, so
drop the second group of tests.

--HG--
extra : rebase_source : 31152f41cf18c5816e6cdea1177f973efedc9fd8