hgweb: fix a crash when using web.archivesubrepos
A matcher is required when enabling the subrepo option on archival.archive(),
because that calls match.narrowmatcher(), which accesses fields on the object.
It's therefore probably a bad idea to default the matcher to None on archive(),
but that's a fix for default.

--HG--
branch : stable
extra : rebase_source : e4c945d4499481926b1b87fd77f522dfee525ccf