largefiles: don't clobber merge action message with user message
The message in the action is used for debugging and should not be the
same as the question presented to the user. Use a different variable
for the user message, so the 'msg' variable already in scope does not
get overwritten.

--HG--
extra : amend_source : b34f3bb0ea069cff21ab4d3c07b02e982dba1667