revbranchcache: move cache writing to the transaction finalizer
Instead of writing the revbranchcache during updatecache (which often happens
too early, before the cache is even populated), let's run it as part of the
transaction finalizer. It still won't be written for read-only operations, but
that's no worse than it is today.

A future commit will remove the actual write that happens in updatecache().

This is also good prep for when all caches get moved into the transaction.

--HG--
extra : rebase_source : e7c3ec7398663117febc4774345d038a13b5faa3