manifest: duplicate call to addrevision()
When we start writing submanifests to their own revlogs, we will not
want to write a new revision for a directory if there were no changes
to it. To prepare for this, duplicate the call to addrevision() and
move them earlier where they can more easily be avoided.

--HG--
extra : rebase_source : eb9cd3fcac9edf7af4a3acd28baa28e6a67c53e4